+++ R/species_params.R
@@ -30,13 +30,11 @@
 #' `calculated_species_params()`. You get all species_params with
 #' `species_params()`.
 #'
-#' If you change given species parameters with `given_species_params<-()` this
-#' will trigger a re-calculation of the calculated species parameters, where
-#' necessary. However if you change species parameters with `species_params<-()`
-#' no recalculation will take place and furthermore your values could be
-#' overwritten by a future recalculation triggered by a call to
-#' `given_species_params<-()` . So in most use cases you will only want to use
-#' `given_species_params<-()`.
+#' When you change species parameters with `species_params<-()`, mizer
+#' automatically detects which parameters you have changed. It records these
+#' changed parameters in `given_species_params` so that they are protected
+#' against being overwritten by future recalculations. It then triggers a
+#' re-calculation of the calculated species parameters.
 #'
 #' There are some species parameters that are used to set up the
 #' size-dependent parameters that are used in the mizer model:
@@ -61,7 +59,7 @@
 #'   using other predation kernel functions.
 #'
 #' When you change one of the above species parameters using
-#' `given_species_params<-()` or `species_params<-()`, the new value will be
+#' `species_params<-()` or `given_species_params<-()`, the new value will be
 #' used to update the corresponding size-dependent rates automatically, unless
 #' you have set those size-dependent rates manually, in which case the
 #' corresponding species parameters will be ignored.
@@ -108,12 +106,9 @@
 #'   `k_vb`, `w_inf` and `t0` as well as the weight-length exponent `b` to
 #'   determine it. This is unreliable and is therefore not recommended.
 #'
-#' Changing these parameters with `species_params<-()` updates the stored
-#' species parameter table and triggers a recalculation via [setParams()].
-#' However they only affect model behaviour if the corresponding downstream
-#' parameters are recalculated rather than kept at explicitly supplied values.
-#' In typical workflows these quantities should therefore be changed via
-#' `given_species_params<-()`.
+#' Changing these parameters with `species_params<-()` will trigger a
+#' recalculation of the downstream parameters, provided they are not protected
+#' by being explicitly given.
 #'
 #' There are other species parameters that are used in tuning the model to
 #' observations:
@@ -149,17 +144,18 @@
 #' @return `species_params()`: Data frame containing all species parameters
 #'   currently stored in the model.
 #'
-#'   `species_params<-()`: Updates the full species parameter table after
-#'   validating it with [validSpeciesParams()] and then recalculating the model
-#'   parameters with [setParams()].
+#'   `species_params<-()`: Updates the `given_species_params` with any
+#'   parameters you have changed, and then recalculates the full species
+#'   parameter table and the model parameters.
 #'
 #'   `given_species_params()`: Data frame containing the species parameter
 #'   values that were supplied explicitly by the user.
 #'
-#'   `given_species_params<-()`: Updates the explicitly supplied species
-#'   parameters after validating them with [validGivenSpeciesParams()] and then
-#'   recalculating the full species parameter table and dependent model
-#'   quantities.
+#'   `given_species_params<-()`: An alternative to `species_params<-()` that
+#'   also triggers a recalculation of other parameters. The only difference is
+#'   that `given_species_params<-()` issues warnings when a parameter is
+#'   changed whose effect is overridden by another parameter that has already
+#'   been given. This is especially useful during interactive use.
 #'
 #'   `calculated_species_params()`: Data frame containing only those species
 #'   parameter entries that are not explicit user input. Columns that would
@@ -230,8 +226,34 @@ species_params.species_params <- function(object, strict = FALSE, ...) {
     if (!all(value$species == object@species_params$species)) {
         stop("The species names in the new species parameter data frame do not match the species names in the model.")
     }
-    object@species_params <- value
-    suppressMessages(setParams(object))
+    
+    # Find what changed compared to old species_params
+    old_sp <- object@species_params
+    given <- object@given_species_params
+    
+    common_cols <- intersect(names(value), names(old_sp))
+    for (col in common_cols) {
+        old_vals <- old_sp[[col]]
+        new_vals <- value[[col]]
+        # which ones changed?
+        changed <- !((old_vals == new_vals) | (is.na(old_vals) & is.na(new_vals)))
+        changed[is.na(changed)] <- TRUE
+        
+        if (any(changed)) {
+            if (!col %in% names(given)) {
+                given[[col]] <- NA
+            }
+            given[[col]][changed] <- new_vals[changed]
+        }
+    }
+    new_cols <- setdiff(names(value), names(old_sp))
+    if (length(new_cols) > 0) {
+        given <- cbind(given, value[new_cols])
+    }
+    
+    object@given_species_params <- given
+    object@species_params <- validSpeciesParams(given)
+    return(suppressMessages(setParams(object)))
 }
 
 #' Test if an object is a species_params object
@@ -372,25 +394,16 @@ check_and_convert_species_params <- function(x) {
 
 #' @export
 print.species_params <- function(x, ...) {
-    cat("An object of class \"species_params\" containing parameters for", nrow(x), "species:\n")
-    core_cols <- c("species", "w_inf", "w_mat", "w_min", "alpha", "erepro")
+    cat("An object of class \"", class(x)[1], "\" containing parameters for ", nrow(x), " species:\n", sep = "")
+    core_cols <- c("species", "w_inf", "w_mat", "h", "ks", "z0", "z_ext")
     cols_to_show <- intersect(core_cols, names(x))
     extra_cols <- setdiff(names(x), core_cols)
 
-    print(as.data.frame(x)[, cols_to_show, drop = FALSE], row.names = FALSE, ...)
-
-    if (length(extra_cols) > 0) {
-        cat("With", length(extra_cols), "other parameters:", paste(extra_cols, collapse = ", "), "\n")
+    if (length(cols_to_show) < length(core_cols) && length(extra_cols) > 0) {
+        num_to_add <- min(length(core_cols) - length(cols_to_show), length(extra_cols))
+        cols_to_show <- c(cols_to_show, extra_cols[1:num_to_add])
+        extra_cols <- extra_cols[-(1:num_to_add)]
     }
-    invisible(x)
-}
-
-#' @export
-print.given_species_params <- function(x, ...) {
-    cat("An object of class \"given_species_params\" containing given parameters for", nrow(x), "species:\n")
-    core_cols <- c("species", "w_inf", "w_mat", "w_min", "alpha", "erepro")
-    cols_to_show <- intersect(core_cols, names(x))
-    extra_cols <- setdiff(names(x), core_cols)
 
     print(as.data.frame(x)[, cols_to_show, drop = FALSE], row.names = FALSE, ...)
 
@@ -461,12 +474,12 @@ given_species_params.data.frame <- function(object, strict = FALSE, ...) {
     }
     sp$species <- species_names
     row.names(sp) <- species_names
-    
+
     # Allow r_max instead of R_max
     if (!("R_max" %in% names(sp)) && "r_max" %in% names(sp)) {
         names(sp)[names(sp) == "r_max"] <- "R_max"
     }
-    
+
     # Convert lengths to weights
     if (all(c("a", "b") %in% names(sp))) {
         sp <- sp %>%
@@ -477,7 +490,7 @@ given_species_params.data.frame <- function(object, strict = FALSE, ...) {
             set_species_param_from_length("w_max", "l_max") %>%
             set_species_param_from_length("w_min", "l_min")
     }
-    
+
     # check w_inf
     if (!("w_inf" %in% names(sp))) {
         if ("w_repro_max" %in% names(sp)) {
@@ -501,7 +514,7 @@ given_species_params.data.frame <- function(object, strict = FALSE, ...) {
             stop("`w_inf` contains non-numeric values.")
         }
     }
-    
+
     # check w_mat
     if ("w_mat" %in% names(sp) && "w_inf" %in% names(sp)) {
         wrong <- !is.na(sp$w_mat) & !is.na(sp$w_inf) & sp$w_mat >= sp$w_inf
@@ -512,44 +525,44 @@ given_species_params.data.frame <- function(object, strict = FALSE, ...) {
                     " I have corrected that by setting it to 25% of `w_inf`.")
             sp$w_mat[wrong] <- sp$w_inf[wrong] / 4
         }
-        
+
         # check w_mat25
         if ("w_mat25" %in% names(sp)) {
             wrong <- !is.na(sp$w_mat) & !is.na(sp$w_mat25) & sp$w_mat25 >= sp$w_mat
             if (any(wrong)) {
-                warning("For the species ", 
+                warning("For the species ",
                         paste(sp$species[wrong], collapse = ", "),
                         " the value for `w_mat25` is not smaller than that of `w_mat`.",
                         " I have corrected that by setting it to NA.")
                 sp$w_mat25[wrong] <- NA
             }
         }
-        
+
         # check w_min
         if ("w_min" %in% names(sp)) {
             wrong <- !is.na(sp$w_min) & !is.na(sp$w_mat) & sp$w_min >= sp$w_mat
             if (any(wrong)) {
                 sp$w_min[wrong] <- pmin(0.001, sp$w_mat[wrong] / 10)
-                warning("For the species ", 
+                warning("For the species ",
                         paste(sp$species[wrong], collapse = ", "),
                         " the value for `w_min` is not smaller than that of `w_mat`.",
                         " I have reduced the values.")
             }
         }
     }
-    
+
     # check w_repro_max
     if ("w_repro_max" %in% names(sp) && "w_mat" %in% names(sp)) {
         wrong <- !is.na(sp$w_repro_max) & !is.na(sp$w_mat) & sp$w_repro_max <= sp$w_mat
         if (any(wrong)) {
-            warning("For the species ", 
+            warning("For the species ",
                     paste(sp$species[wrong], collapse = ", "),
                     " the value for `w_repro_max` is smaller than that of `w_mat`.",
                     " I have corrected that by setting it to 4 times `w_mat.")
             sp$w_repro_max[wrong] <- 4 * sp$w_mat[wrong]
         }
     }
-    
+
     class(sp) <- c("given_species_params", "species_params", setdiff(class(sp), c("given_species_params", "species_params")))
     check_and_convert_species_params(sp)
 }
@@ -644,6 +657,9 @@ calculated_species_params <- function(params) {
     # Removing columns that only contain NAs
     calculated <- calculated %>%
         select(where(~ !all(is.na(.))))
+    
+    calculated$species <- params@species_params$species
+    calculated <- calculated[, c("species", setdiff(names(calculated), "species")), drop = FALSE]
 
     return(calculated)
 }